Add alias types

+++ src/codegen.cpp
@@ -836,7 +836,7 @@ void codegen(SymbolTable& symbol_table, FILE* file, bool is_libc_mode)
 			if (!func.return_type_index.has_value())
 				log_error(func.ast[func.ast_node_root], "Main function missing return type");
 
-			if (symbol_table.types[func.return_type_index.value()].name != "int")
+			if (symbol_table.types[func.return_type_index.value()].name != "u64")
 				log_error(func.ast[func.ast_node_root], "Main function defined with wrong return type");
 
 			main_defined = true;
